Move get_data macro to submacros module
